pyproject.toml
[project]
name = "repl-nix-arctecfox"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "sdv>=1.17.3",
    "streamlit>=1.41.1",
]
